PCINetDevices: use sysfs to enumerate ethernet devices
The `PCINetDevice` constructor already uses sysfs to initialize
itself so this would not introduce a behavioural change. It
does however allow use of the PCI library with the `netdevsim`
driver which would help in downstream testing.

Use existence of `sriov_numvfs` as indication of SR-IOV support
instead of `sriov_totalvfs` to allow testing with `netdevsim`.

Handle non-existence of `sriov_totalvfs` by detecting the
`netdevsim` driver and returning a fixed number for totalvfs.